+++ scripts/DBUpdate-to-6.postgresql.sql
@@ -12,4 +12,5 @@ CREATE TABLE dynamic_field_obj_id_name (
     PRIMARY KEY(object_id),
     CONSTRAINT dynamic_field_object_name UNIQUE (object_name, object_type)
 );
+CREATE INDEX dynamic_field_value_search_text ON dynamic_field_value (field_id, value_text);
 SET standard_conforming_strings TO ON;